# RedLab/http/E01_openapi.json
{"openapi":"3.1.0","info":{"title":"redlab reporting/billing service","version":"0.1.0"},"paths":{"/reports":{"get":{"summary":"List Reports","operationId":"list_reports_reports_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ReportOut"},"type":"array","title":"Response List Reports Reports Get"}}}}}},"post":{"summary":"Create Report","operationId":"create_report_reports_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/{report_id}":{"get":{"summary":"Get Report","operationId":"get_report_reports__report_id__get","parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Report","operationId":"delete_report_reports__report_id__delete","parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/billing/tenants":{"get":{"summary":"List Billing Tenants","description":"Normal billing read. Stage 0 observes tenant_ids through the app. Which\ntenants are the real 6 (tenants-of-record) is NOT revealed here — that needs\nthe manifest (Stage 2/3). Derivation clues are human-owned.","operationId":"list_billing_tenants_billing_tenants_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BillingTenantOut"},"type":"array","title":"Response List Billing Tenants Billing Tenants Get"}}}}}}},"/reports/import":{"post":{"tags":["import"],"summary":"Create Import","operationId":"create_import_reports_import_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportJobOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/import/{job_id}":{"get":{"tags":["import"],"summary":"Get Import","operationId":"get_import_reports_import__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportJobOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/import/{job_id}/transition":{"post":{"tags":["import"],"summary":"Transition","operationId":"transition_reports_import__job_id__transition_post","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransitionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportJobOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/healthz":{"get":{"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BillingTenantOut":{"properties":{"tenant_id":{"type":"string","title":"Tenant Id"},"plan":{"type":"string","title":"Plan"},"mrr_krw":{"type":"integer","title":"Mrr Krw"},"status":{"type":"string","title":"Status"}},"type":"object","required":["tenant_id","plan","mrr_krw","status"],"title":"BillingTenantOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImportJobOut":{"properties":{"id":{"type":"integer","title":"Id"},"report_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Report Id"},"source_url":{"type":"string","title":"Source Url"},"state":{"type":"string","title":"State"},"detail":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Detail"}},"type":"object","required":["id","source_url","state"],"title":"ImportJobOut"},"ImportRequest":{"properties":{"source_url":{"type":"string","title":"Source Url"},"method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Method","default":"GET"},"headers":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Headers"}},"type":"object","required":["source_url"],"title":"ImportRequest"},"ReportCreate":{"properties":{"title":{"type":"string","title":"Title"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"}},"type":"object","required":["title"],"title":"ReportCreate"},"ReportOut":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"source_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Url"},"status":{"type":"string","title":"Status"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","title","status","created_at"],"title":"ReportOut"},"TransitionRequest":{"properties":{"to":{"type":"string","title":"To"}},"type":"object","required":["to"],"title":"TransitionRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}